Do not treat // within string as C++ style comment
The use of // to indicate a comment that extends to the end of line
was a recent change. It has been noticed that this feature does not
allow the use of URLs in calendar lines.

Change the parsing of // to only start a comment at the beginning of
a line or when preceded by white space.